:hammer: refactor: generator domain 패키지명 변경
