+++ src/main/java/tipitapi/drawmytoday/domain/generator/api/dalle/dto/CreateDallEImageRequest.java
@@ -1,0 +1,25 @@
+package tipitapi.drawmytoday.domain.generator.api.dalle.dto;
+
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@NoArgsConstructor
+@AllArgsConstructor(access = AccessLevel.PRIVATE)
+public class CreateDallEImageRequest {
+
+    private String prompt;
+    private int n;
+    private String size;
+    private String response_format;
+
+    public static CreateDallEImageRequest withBase64(String prompt) {
+        return new CreateDallEImageRequest(prompt, 1, "1024x1024", "b64_json");
+    }
+
+    public static CreateDallEImageRequest withUrl(String prompt) {
+        return new CreateDallEImageRequest(prompt, 1, "512x512", "url");
+    }
+}
