+++ src/main/java/tipitapi/drawmytoday/domain/generator/api/dalle/exception/DallEPolicyViolationException.java
@@ -1,0 +1,11 @@
+package tipitapi.drawmytoday.domain.generator.api.dalle.exception;
+
+import tipitapi.drawmytoday.common.exception.ErrorCode;
+import tipitapi.drawmytoday.domain.generator.exception.ImageGeneratorException;
+
+public class DallEPolicyViolationException extends ImageGeneratorException {
+
+    public DallEPolicyViolationException(Throwable throwable) {
+        super(ErrorCode.DALLE_CONTENT_POLICY_VIOLATION, throwable);
+    }
+}
