+++ src/main/java/tipitapi/drawmytoday/domain/generator/api/gpt/domain/Message.java
@@ -1,0 +1,17 @@
+package tipitapi.drawmytoday.domain.generator.api.gpt.domain;
+
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@NoArgsConstructor(force = true)
+public class Message {
+
+    private final ChatCompletionsRole role;
+    private final String content;
+
+    public Message(ChatCompletionsRole role, String content) {
+        this.role = role;
+        this.content = content;
+    }
+}
