+++ src/main/java/tipitapi/drawmytoday/domain/generator/api/gpt/dto/ChoiceResponse.java
@@ -1,0 +1,23 @@
+package tipitapi.drawmytoday.domain.generator.api.gpt.dto;
+
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import com.fasterxml.jackson.databind.PropertyNamingStrategies;
+import com.fasterxml.jackson.databind.annotation.JsonNaming;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import tipitapi.drawmytoday.domain.generator.api.gpt.domain.Message;
+
+@Getter
+@NoArgsConstructor
+@AllArgsConstructor
+@JsonIgnoreProperties(ignoreUnknown = true)
+@JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
+public class ChoiceResponse {
+
+    private String firstReason;
+    private Long index;
+    private Message message;
+
+
+}
