+++ src/main/java/tipitapi/drawmytoday/domain/generator/api/gpt/dto/GptChatCompletionsResponse.java
@@ -1,0 +1,24 @@
+package tipitapi.drawmytoday.domain.generator.api.gpt.dto;
+
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import com.fasterxml.jackson.databind.PropertyNamingStrategies;
+import com.fasterxml.jackson.databind.annotation.JsonNaming;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+/**
+ * https://platform.openai.com/docs/guides/text-generation/chat-completions-response-format
+ */
+
+@Getter
+@NoArgsConstructor
+@AllArgsConstructor
+@JsonIgnoreProperties(ignoreUnknown = true)
+@JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
+public class GptChatCompletionsResponse {
+
+    private String id;
+    private ChoiceResponse[] choices;
+
+}
