--- src/main/java/tipitapi/drawmytoday/domain/generator/domain/gpt/domain/ChatCompletionsRole.java
@@ -1,15 +1,0 @@
-package tipitapi.drawmytoday.domain.generator.domain.gpt.domain;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-
-public enum ChatCompletionsRole {
-    @JsonProperty("system")
-    system,
-    @JsonProperty("user")
-    user,
-    @JsonProperty("assistant")
-    assistant;
-
-    ChatCompletionsRole() {
-    }
-}
