--- src/main/java/tipitapi/drawmytoday/domain/generator/domain/karlo/dto/KarloImageUrlResponse.java
@@ -1,20 +1,0 @@
-package tipitapi.drawmytoday.domain.generator.domain.karlo.dto;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-import com.fasterxml.jackson.databind.PropertyNamingStrategies;
-import com.fasterxml.jackson.databind.annotation.JsonNaming;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor
-@AllArgsConstructor
-@JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
-public class KarloImageUrlResponse {
-
-    private String id;
-    private Long seed;
-    @JsonProperty("image")
-    private String imageUrl;
-}
